Fixed imageviewer and other programs by making width etc long long.

+++ Convert_16-8bit/convert_16-8bit.cpp
@@ -36,7 +36,7 @@ unsigned short *p_16;
 int main(int argc, char**argv)
 {
 	int x, y, z, nbits, thresh;
-	int width, height, depth, xysize;
+	long long width, height, depth, xysize;
 	bool use_compression = true;
 
 	if (argc != 5) {
